[Search] Unify search plugin step 2 — Search Interceptor (#97845)
Unify search interceptor
Deprecate TimeoutErrorMode.UPGRADE
