--- docs/development/plugins/data/public/kibana-plugin-plugins-data-public.searchinterceptor.deps.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [kibana-plugin-plugins-data-public](./kibana-plugin-plugins-data-public.md) &gt; [SearchInterceptor](./kibana-plugin-plugins-data-public.searchinterceptor.md) &gt; [deps](./kibana-plugin-plugins-data-public.searchinterceptor.deps.md)
-
-## SearchInterceptor.deps property
-
-<b>Signature:</b>
-
-```typescript
-protected readonly deps: SearchInterceptorDeps;
-```
