--- docs/development/plugins/data/public/kibana-plugin-plugins-data-public.searchinterceptor.gettimeoutmode.md
@@ -1,15 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [kibana-plugin-plugins-data-public](./kibana-plugin-plugins-data-public.md) &gt; [SearchInterceptor](./kibana-plugin-plugins-data-public.searchinterceptor.md) &gt; [getTimeoutMode](./kibana-plugin-plugins-data-public.searchinterceptor.gettimeoutmode.md)
-
-## SearchInterceptor.getTimeoutMode() method
-
-<b>Signature:</b>
-
-```typescript
-protected getTimeoutMode(): TimeoutErrorMode;
-```
-<b>Returns:</b>
-
-`TimeoutErrorMode`
-
