--- docs/development/plugins/data/public/kibana-plugin-plugins-data-public.searchinterceptor.handlesearcherror.md
@@ -1,24 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [kibana-plugin-plugins-data-public](./kibana-plugin-plugins-data-public.md) &gt; [SearchInterceptor](./kibana-plugin-plugins-data-public.searchinterceptor.md) &gt; [handleSearchError](./kibana-plugin-plugins-data-public.searchinterceptor.handlesearcherror.md)
-
-## SearchInterceptor.handleSearchError() method
-
-<b>Signature:</b>
-
-```typescript
-protected handleSearchError(e: KibanaServerError | AbortError, options?: ISearchOptions, isTimeout?: boolean): Error;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  e | <code>KibanaServerError &#124; AbortError</code> |  |
-|  options | <code>ISearchOptions</code> |  |
-|  isTimeout | <code>boolean</code> |  |
-
-<b>Returns:</b>
-
-`Error`
-
