+++ docs/development/plugins/data/public/kibana-plugin-plugins-data-public.searchinterceptor.search.md
@@ -9,19 +9,19 @@ Searches using the given `search` method. Overrides the `AbortSignal` with one t
 <b>Signature:</b>
 
 ```typescript
-search(request: IKibanaSearchRequest, options?: ISearchOptions): Observable<IKibanaSearchResponse>;
+search({ id, ...request }: IKibanaSearchRequest, options?: IAsyncSearchOptions): import("rxjs").Observable<IKibanaSearchResponse<any>>;
 ```
 
 ## Parameters
 
 |  Parameter | Type | Description |
 |  --- | --- | --- |
-|  request | <code>IKibanaSearchRequest</code> |  |
-|  options | <code>ISearchOptions</code> |  |
+|  { id, ...request } | <code>IKibanaSearchRequest</code> |  |
+|  options | <code>IAsyncSearchOptions</code> |  |
 
 <b>Returns:</b>
 
-`Observable<IKibanaSearchResponse>`
+`import("rxjs").Observable<IKibanaSearchResponse<any>>`
 
 `Observable` emitting the search response or an error.
