+++ docs/development/plugins/data/public/kibana-plugin-plugins-data-public.searchinterceptor.stop.md
@@ -1,0 +1,15 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [kibana-plugin-plugins-data-public](./kibana-plugin-plugins-data-public.md) &gt; [SearchInterceptor](./kibana-plugin-plugins-data-public.searchinterceptor.md) &gt; [stop](./kibana-plugin-plugins-data-public.searchinterceptor.stop.md)
+
+## SearchInterceptor.stop() method
+
+<b>Signature:</b>
+
+```typescript
+stop(): void;
+```
+<b>Returns:</b>
+
+`void`
+
