+++ src/plugins/data/public/search/search_interceptor/search_response_cache.test.ts
@@ -1,8 +1,9 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
 import { interval, Observable, of, throwError } from 'rxjs';
@@ -14,7 +15,7 @@ import { SearchResponseCache } from './search_response_cache';
 describe('SearchResponseCache', () => {
   let cache: SearchResponseCache;
   let searchAbortController: SearchAbortController;
-  const r: Array<IKibanaSearchResponse<any>> = [
+  const r: IKibanaSearchResponse[] = [
     {
       isPartial: true,
       isRunning: true,
