+++ src/plugins/data/public/search/search_interceptor/search_response_cache.ts
@@ -1,21 +1,22 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
 import { Observable, Subscription } from 'rxjs';
-import { IKibanaSearchResponse, isErrorResponse } from '../../../../../src/plugins/data/public';
 import { SearchAbortController } from './search_abort_controller';
+import { IKibanaSearchResponse, isErrorResponse } from '../../../common';
 
 interface ResponseCacheItem {
-  response$: Observable<IKibanaSearchResponse<any>>;
+  response$: Observable<IKibanaSearchResponse>;
   searchAbortController: SearchAbortController;
 }
 
 interface ResponseCacheItemInternal {
-  response$: Observable<IKibanaSearchResponse<any>>;
+  response$: Observable<IKibanaSearchResponse>;
   searchAbortController: SearchAbortController;
   size: number;
   subs: Subscription;
