+++ src/plugins/data/public/search/types.ts
@@ -7,7 +7,6 @@
  */
 
 import { PackageInfo } from 'kibana/server';
-import { ISearchInterceptor } from './search_interceptor';
 import { SearchUsageCollector } from './collectors';
 import { AggsSetup, AggsSetupDependencies, AggsStartDependencies, AggsStart } from './aggs';
 import { ISearchGeneric, ISearchStartSearchSource } from '../../common/search';
@@ -16,10 +15,6 @@ import { UsageCollectionSetup } from '../../../usage_collection/public';
 import { ISessionsClient, ISessionService } from './session';
 
 export { ISearchStartSearchSource, SearchUsageCollector };
-
-export interface SearchEnhancements {
-  searchInterceptor: ISearchInterceptor;
-}
 
 /**
  * The setup contract exposed by the Search plugin exposes the search strategy extension
@@ -38,10 +33,6 @@ export interface ISearchSetup {
    * {@link ISessionsClient}
    */
   sessionsClient: ISessionsClient;
-  /**
-   * @internal
-   */
-  __enhance: (enhancements: SearchEnhancements) => void;
 }
 
 /**
