+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -1419,8 +1419,6 @@
     "data.search.timeoutIncreaseSetting": "クエリがタイムアウトしました。検索タイムアウト詳細設定で実行時間を延長します。",
     "data.search.timeoutIncreaseSettingActionText": "設定を編集",
     "data.search.unableToGetSavedQueryToastTitle": "保存したクエリ {savedQueryId} を読み込めません",
-    "data.search.upgradeLicense": "クエリがタイムアウトしました。無料のベーシックティアではクエリがタイムアウトすることはありません。",
-    "data.search.upgradeLicenseActionText": "今すぐアップグレード",
     "data.searchSessionIndicator.noCapability": "検索セッションを作成するアクセス権がありません。",
     "data.triggers.applyFilterDescription": "Kibanaフィルターが適用されるとき。単一の値または範囲フィルターにすることができます。",
     "data.triggers.applyFilterTitle": "フィルターを適用",
@@ -1441,6 +1439,8 @@
     "data.kueryAutocomplete.lessThanOrEqualOperatorDescription.lessThanOrEqualToText": "より小さいまたは等しい",
     "data.kueryAutocomplete.orOperatorDescription": "{oneOrMoreArguments} が true であることを条件とする",
     "data.kueryAutocomplete.orOperatorDescription.oneOrMoreArgumentsText": "1つ以上の引数",
+    "data.advancedSettings.searchTimeout": "検索タイムアウト",
+    "data.advancedSettings.searchTimeoutDesc": "検索セッションの最大タイムアウトを変更するか、0に設定してタイムアウトを無効にすると、クエリは完了するまで実行されます。",
     "devTools.badge.readOnly.text": "読み取り専用",
     "devTools.badge.readOnly.tooltip": "を保存できませんでした",
     "devTools.devToolsTitle": "開発ツール",
@@ -7250,8 +7250,6 @@
     "xpack.dashboardMode.uiSettings.dashboardsOnlyRolesDeprecation": "この設定はサポートが終了し、Kibana 8.0 では削除されます。",
     "xpack.dashboardMode.uiSettings.dashboardsOnlyRolesDescription": "ダッシュボード表示専用モードのロールです",
     "xpack.dashboardMode.uiSettings.dashboardsOnlyRolesTitle": "ダッシュボード専用ロール",
-    "xpack.data.advancedSettings.searchTimeout": "検索タイムアウト",
-    "xpack.data.advancedSettings.searchTimeoutDesc": "検索セッションの最大タイムアウトを変更するか、0に設定してタイムアウトを無効にすると、クエリは完了するまで実行されます。",
     "xpack.data.mgmt.searchSessions.actionDelete": "削除",
     "xpack.data.mgmt.searchSessions.actionExtend": "延長",
     "xpack.data.mgmt.searchSessions.actions.tooltip.moreActions": "さらにアクションを表示",
