+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -1427,8 +1427,6 @@
     "data.search.timeoutIncreaseSetting": "您的查询已超时。使用搜索超时高级设置延长运行时间。",
     "data.search.timeoutIncreaseSettingActionText": "编辑设置",
     "data.search.unableToGetSavedQueryToastTitle": "无法加载已保存查询 {savedQueryId}",
-    "data.search.upgradeLicense": "您的查询已超时。使用我们免费的基本级许可，您的查询永不会超时。",
-    "data.search.upgradeLicenseActionText": "立即升级",
     "data.searchSessionIndicator.noCapability": "您无权创建搜索会话。",
     "data.triggers.applyFilterDescription": "应用 kibana 筛选时。可能是单个值或范围筛选。",
     "data.triggers.applyFilterTitle": "应用筛选",
@@ -1449,6 +1447,8 @@
     "data.kueryAutocomplete.lessThanOrEqualOperatorDescription.lessThanOrEqualToText": "小于或等于",
     "data.kueryAutocomplete.orOperatorDescription": "需要{oneOrMoreArguments}为 true",
     "data.kueryAutocomplete.orOperatorDescription.oneOrMoreArgumentsText": "一个或多个参数",
+    "data.advancedSettings.searchTimeout": "搜索超时",
+    "data.advancedSettings.searchTimeoutDesc": "更改搜索会话的最大超时值，或设置为 0 以禁用超时，让查询运行至结束。",
     "devTools.badge.readOnly.text": "只读",
     "devTools.badge.readOnly.tooltip": "无法保存",
     "devTools.devToolsTitle": "开发工具",
@@ -7313,8 +7313,6 @@
     "xpack.dashboardMode.uiSettings.dashboardsOnlyRolesDeprecation": "此设置已过时，将在 Kibana 8.0 中移除。",
     "xpack.dashboardMode.uiSettings.dashboardsOnlyRolesDescription": "属于“仅查看仪表板”模式的角色",
     "xpack.dashboardMode.uiSettings.dashboardsOnlyRolesTitle": "仅限仪表板的角色",
-    "xpack.data.advancedSettings.searchTimeout": "搜索超时",
-    "xpack.data.advancedSettings.searchTimeoutDesc": "更改搜索会话的最大超时值，或设置为 0 以禁用超时，让查询运行至结束。",
     "xpack.data.mgmt.searchSessions.actionDelete": "删除",
     "xpack.data.mgmt.searchSessions.actionExtend": "延长",
     "xpack.data.mgmt.searchSessions.actions.tooltip.moreActions": "更多操作",
